65 - Enhancement - Ajuste de estrutura do projeto

+++ src/main/java/br/com/fiap/fastfood/api/adapters/driven/infrastructure/repository/adapter/OrderRepositoryAdapterImpl.java
@@ -1,16 +1,15 @@
 package br.com.fiap.fastfood.api.adapters.driven.infrastructure.repository.adapter;
 
 import br.com.fiap.fastfood.api.adapters.driven.infrastructure.entity.order.OrderEntity;
-import br.com.fiap.fastfood.api.adapters.driven.infrastructure.mapper.InvoiceMapper;
-import br.com.fiap.fastfood.api.adapters.driven.infrastructure.mapper.InvoiceMapperImpl;
-import br.com.fiap.fastfood.api.adapters.driven.infrastructure.mapper.OrderMapper;
-import br.com.fiap.fastfood.api.adapters.driven.infrastructure.mapper.OrderMapperImpl;
+import br.com.fiap.fastfood.api.adapters.driven.infrastructure.mapper.InvoiceMapperInfra;
+import br.com.fiap.fastfood.api.adapters.driven.infrastructure.mapper.InvoiceMapperInfraImpl;
+import br.com.fiap.fastfood.api.adapters.driven.infrastructure.mapper.OrderMapperInfra;
+import br.com.fiap.fastfood.api.adapters.driven.infrastructure.mapper.OrderMapperInfraImpl;
 import br.com.fiap.fastfood.api.adapters.driven.infrastructure.repository.order.OrderRepository;
-import br.com.fiap.fastfood.api.core.application.ports.repository.OrderRepositoryPort;
-import br.com.fiap.fastfood.api.core.domain.model.invoice.Invoice;
-import br.com.fiap.fastfood.api.core.domain.model.order.Order;
+import br.com.fiap.fastfood.api.core.application.dto.invoice.InvoiceDTO;
+import br.com.fiap.fastfood.api.core.application.dto.order.OrderDTO;
+import br.com.fiap.fastfood.api.core.application.port.repository.OrderRepositoryPort;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -21,48 +20,51 @@ import org.springframework.stereotype.Component;
 public class OrderRepositoryAdapterImpl implements OrderRepositoryPort {
 
   private final OrderRepository orderRepository;
-  private final OrderMapper orderMapper;
-  private final InvoiceMapper invoiceMapper;
+  private final OrderMapperInfra orderMapperInfra;
+  private final InvoiceMapperInfra invoiceMapperInfra;
 
   @Autowired
   public OrderRepositoryAdapterImpl(OrderRepository orderRepository) {
     this.orderRepository = orderRepository;
-    this.orderMapper = new OrderMapperImpl();
-    this.invoiceMapper = new InvoiceMapperImpl();
+    this.orderMapperInfra = new OrderMapperInfraImpl();
+    this.invoiceMapperInfra = new InvoiceMapperInfraImpl();
   }
 
   @Override
-  public Optional<Order> findById(Long identifier) {
+  public Optional<OrderDTO> findById(Long identifier) {
     Optional<OrderEntity> orderEntityOptional = orderRepository.findById(identifier);
-    Optional<Order> orderOpt = orderEntityOptional.map(orderMapper::toDomain);
+    Optional<OrderDTO> orderOpt = orderEntityOptional.map(orderMapperInfra::toDTO);
     if (orderEntityOptional.isPresent() && orderOpt.isPresent()) {
       OrderEntity entity = orderEntityOptional.get();
-      Order order = orderOpt.get();
-      List<Invoice> invoices = Optional.ofNullable(entity.getOrderInvoices()).orElse(Collections.emptyList()).stream().map(i -> {
-        Invoice invoice = invoiceMapper.toDomain(i);
-        invoiceMapper.mapStateImpl(i.getState(), invoice);
-        return invoice;
-      }).collect(
-          Collectors.toList());
-      order.setInvoices(invoices);
-      order.setState(orderMapper.mapStateImpl(entity
-          .getState(), order));
+      OrderDTO orderDTO = orderOpt.get();
+      List<InvoiceDTO> invoices = Optional.ofNullable(entity.getOrderInvoices())
+          .orElse(new ArrayList<>()).stream().map(
+              invoiceMapperInfra::toDTO).collect(
+              Collectors.toList());
+      orderDTO.setInvoices(invoices);
     }
     return orderOpt;
   }
 
   @Override
-  public Order save(Order data) {
-    OrderEntity entity = orderMapper.toEntity(data);
+  public OrderDTO save(OrderDTO data) {
+    OrderEntity entity = orderMapperInfra.toEntity(data);
     OrderEntity persistedEntity = orderRepository.save(entity);
-    Order order = orderMapper.toDomain(persistedEntity);
-    order.setState(orderMapper.mapStateImpl(persistedEntity.getState(), order));
-    return order;
+    return orderMapperInfra.toDTO(persistedEntity);
   }
 
   @Override
   public boolean delete(Long identifier) {
+    Optional<OrderEntity> orderEntityOpt = orderRepository.findById(identifier);
+    if (orderEntityOpt.isEmpty()) {
+      return false;
+    }
     orderRepository.deleteById(identifier);
     return true;
+  }
+
+  @Override
+  public List<OrderDTO> findAll() {
+    return orderRepository.findAll().stream().map(orderMapperInfra::toDTO).collect(Collectors.toList());
   }
 }
